implemented jwt for authentication and protected all apis with jwt

+++ controllers/constactControllers.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 
 const createContact = asyncHandler(async (req, res) => {
-    console.log("The request body is",req.body)
-    const {name, email, phone} = req.body;
-    if(!name || !email || !phone){
+    console.log("The request body is", req.body)
+    const { name, email, phone } = req.body;
+    if (!name || !email || !phone) {
         res.status(400);
         throw new Error("All fields are mandotory")
     }
@@ -19,6 +19,7 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     })
     res.status(201).json(contact);
 });
@@ -26,7 +27,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
-    if(!contact) {
+    if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
@@ -36,26 +37,34 @@ const getContact = asyncHandler(async (req, res) => {
 
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
-    if(!contact) {
+    if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other users")
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
-        req.params.id, 
+        req.params.id,
         req.body,
-        { new: true}
-        );
+        { new: true }
+    );
 
     res.status(200).json(updatedContact);
 });
 
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
-    if(!contact) {
+    if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
-    await Contact.deleteOne({_id: req.params.id});
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other users")
+    }
+    await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
 });
